+++ pkg/utility.go
@@ -1,4 +1,4 @@
-package utility
+package kln
 
 import (
 	"fmt"
